# perf: speed up network load times