--- a/package.json
+++ b/package.json
@@ -37,7 +37,7 @@
     "@vue/runtime-dom": "3.4.21",
     "rollup-plugin-gzip": "^3.1.0",
     "vite-plugin-static-copy": "^0.17.0",
-    "vite-plugin-treat-umd-as-commonjs": "0.1.3",
+    "vite-plugin-treat-umd-as-commonjs": "0.1.4",
     "vue": "3.4.21"
   },
   "peerDependencies": {