--- a/packages/design-system/src/components/OcDatepicker/OcDatepicker.spec.ts
+++ b/packages/design-system/src/components/OcDatepicker/OcDatepicker.spec.ts
@@ -1,5 +1,10 @@
+import { defineComponent } from 'vue'
 import Datepicker from './OcDatepicker.vue'
 import { mount } from 'web-test-helpers'
+
+const DatePickerComponent = defineComponent({
+  template: '<div id="foo"><slot></slot></div>'
+})
 
 describe('OcDatePicker', () => {
   it('renders default scoped slot', () => {
@@ -8,9 +13,11 @@ describe('OcDatePicker', () => {
       slots: { default: slotDefault },
       props: { value: null },
       global: {
-        renderStubDefaultSlot: true
+        renderStubDefaultSlot: true,
+        stubs: { DatePicker: DatePickerComponent }
       }
     })
+
     expect(wrapper.find('#default-slot').exists()).toBeTruthy()
   })
 })