--- a/packages/web-pkg/src/components/index.ts
+++ b/packages/web-pkg/src/components/index.ts
@@ -10,6 +10,7 @@ export * from './SideBar'
 export * from './Search'
 export * from './Spaces'
 
+export { default as AppBanner } from './AppBanner.vue'
 export { default as AppLoadingSpinner } from './AppLoadingSpinner.vue'
 export { default as AppTopBar } from './AppTopBar.vue'
 export { default as BatchActions } from './BatchActions.vue'