--- a/packages/web-runtime/src/defaults/index.ts
+++ b/packages/web-runtime/src/defaults/index.ts
@@ -1,7 +1,6 @@
 import merge from 'lodash-es/merge'
 import App from '../App.vue'
 import missingOrInvalidConfigPage from '../pages/missingOrInvalidConfig.vue'
-import { coreTranslations, clientTranslations, pkgTranslations, odsTranslations } from './json'
 
 // fontawesome-free attributions console message
 import '@fortawesome/fontawesome-free/attribution'
@@ -10,10 +9,11 @@ export { default as DesignSystem } from 'design-system'
 export * from './languages'
 
 export const pages = { success: App, failure: missingOrInvalidConfigPage }
-export const translations = merge(
-  {},
-  coreTranslations,
-  clientTranslations,
-  pkgTranslations,
-  odsTranslations
-)
+
+export const loadTranslations = async () => {
+  const { coreTranslations, clientTranslations, pkgTranslations, odsTranslations } = await import(
+    './json'
+  )
+
+  return merge({}, coreTranslations, clientTranslations, pkgTranslations, odsTranslations)
+}